projects/Angel_Cloud-snapshots/shanebrain-core/requirements.txt
# =============================================================================
# ShaneBrain Core - Python Dependencies
# =============================================================================
# Install with: pip install -r requirements.txt
# =============================================================================

# LangChain ecosystem
langchain>=0.1.0
langchain-community>=0.0.20
langchain-ollama>=0.1.0

# Vector database
weaviate-client>=4.4.0

# MongoDB
pymongo>=4.6.0

# Ollama Python client
ollama>=0.1.0

# HTTP client (for health checks)
requests>=2.31.0

# Environment variables
python-dotenv>=1.0.0

# Async support
aiohttp>=3.9.0

# Rich terminal output (for Angel Cloud CLI)
rich>=13.7.0

# Prompt toolkit (for interactive CLI)
prompt_toolkit>=3.0.43
